# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: M&S Delivery System API
  version: 1.0.0
  description: Official REST API specification for M&S Delivery System (FastAPI backend on Vercel).
servers:
  - url: /api/v1
    description: Current API v1 base path
paths:
...
  /orders/preview-fee:
    post:
      summary: Calculate estimated delivery fee and distance
      operationId: previewDeliveryFee
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeePreviewRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeePreviewResponse'
  /orders:
    get:
      summary: List orders (Customer gets own, Admin gets all, Rider gets assigned)
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: store_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Place a new COD order
      operationId: createOrder
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{order_id}:
    get:
      summary: Get order details
      operationId: getOrderDetail
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
  /orders/{order_id}/cancel:
    post:
      summary: Cancel an order (Customer if pending, Admin anytime with reason)
      operationId: cancelOrder
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{order_id}/assign:
    post:
      summary: Assign order to a rider (Admin only)
      operationId: assignOrderToRider
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignOrderRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
...
  /riders:
    get:
      summary: List all riders with statuses (Admin only)
      operationId: listRiders
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiderProfileResponse'
    post:
      summary: Create a rider account (Admin only)
      operationId: createRiderAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRiderRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderProfileResponse'
  /riders/status:
    put:
      summary: Update rider availability status (available, busy, offline)
      operationId: updateRiderStatus
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderStatusUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderProfileResponse'
...
  /reports/sales:
    get:
      summary: Generate sales & order report
      operationId: getSalesReport
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: store_id
          in: query
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesReportResponse'
            text/csv:
              schema:
                type: string
  /reports/riders:
    get:
      summary: Generate rider performance report
      operationId: getRiderReport
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderReportResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    OrderStatus:
      type: string
      enum: [pending, confirmed, assigned, picked_up, on_the_way, delivered, cancelled]
    PaymentStatus:
      type: string
      enum: [unpaid, paid]
    RiderStatus:
      type: string
      enum: [available, busy, offline]
...
    FeePreviewRequest:
      type: object
      required: [store_id, delivery_latitude, delivery_longitude]
      properties:
        store_id:
          type: string
          format: uuid
        delivery_latitude:
          type: number
        delivery_longitude:
          type: number
    FeePreviewResponse:
      type: object
      required: [base_fee, distance_km, per_km_fee, total_delivery_fee, estimated_duration_minutes]
      properties:
        base_fee:
          type: number
        distance_km:
          type: number
        per_km_fee:
          type: number
        total_delivery_fee:
          type: number
        estimated_duration_minutes:
          type: integer
    OrderItemCreate:
      type: object
      required: [menu_item_id, quantity]
      properties:
        menu_item_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        special_instructions:
          type: string
    OrderCreateRequest:
      type: object
      required: [store_id, address_id, items]
      properties:
        store_id:
          type: string
          format: uuid
        address_id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        notes:
          type: string
    OrderItem:
      type: object
      required: [id, menu_item_id, item_name, unit_price, quantity, subtotal]
      properties:
        id:
          type: string
          format: uuid
        menu_item_id:
          type: string
          format: uuid
        item_name:
          type: string
        unit_price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
        special_instructions:
          type: string
    Payment:
      type: object
      required: [id, method, amount, status]
      properties:
        id:
          type: string
          format: uuid
        method:
          type: string
          example: cash_on_delivery
        amount:
          type: number
        status:
          $ref: '#/components/schemas/PaymentStatus'
        collected_at:
          type: string
          format: date-time
    Delivery:
      type: object
      required: [id, order_id, status]
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        rider_id:
          type: string
          format: uuid
        rider_name:
          type: string
        rider_phone:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        pickup_time:
          type: string
          format: date-time
        delivered_time:
          type: string
          format: date-time
        last_latitude:
          type: number
        last_longitude:
          type: number
        last_location_time:
          type: string
          format: date-time
    Order:
      type: object
      required: [id, order_number, customer_id, store_id, store_name, status, subtotal, delivery_fee, total_amount, payment_status, created_at]
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
        customer_phone:
          type: string
        store_id:
          type: string
          format: uuid
        store_name:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotal:
          type: number
        delivery_fee:
          type: number
        total_amount:
          type: number
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        delivery_address:
          type: string
        delivery_latitude:
          type: number
        delivery_longitude:
          type: number
        created_at:
          type: string
          format: date-time
    OrderDetail:
      type: object
      required: [order, items, payment]
      properties:
        order:
          $ref: '#/components/schemas/Order'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payment:
          $ref: '#/components/schemas/Payment'
        delivery:
          $ref: '#/components/schemas/Delivery'
    CancelOrderRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    AssignOrderRequest:
      type: object
      required: [rider_id]
      properties:
        rider_id:
          type: string
          format: uuid
...
    CreateRiderRequest:
      type: object
      required: [email, password, full_name, phone, vehicle_type, plate_number]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        phone:
          type: string
        vehicle_type:
          type: string
          example: Motorcycle
        plate_number:
          type: string
    RiderProfileResponse:
      type: object
      required: [id, user_id, full_name, email, phone, status, vehicle_type, plate_number]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/RiderStatus'
        vehicle_type:
          type: string
        plate_number:
          type: string
        active_delivery_id:
          type: string
          format: uuid
    RiderStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/RiderStatus'
...
    SalesReportResponse:
      type: object
      required: [total_orders, total_sales, total_delivery_fees, orders]
      properties:
        total_orders:
          type: integer
        total_sales:
          type: number
        total_delivery_fees:
          type: number
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    RiderReportResponse:
      type: object
      required: [riders]
      properties:
        riders:
          type: array
          items:
            type: object
            properties:
              rider_id:
                type: string
              rider_name:
                type: string
              completed_deliveries:
                type: integer
              total_distance_km:
                type: number
              average_delivery_time_minutes:
                type: number
...